moved  config_page to a package
